requirements.txt
Django==4.1.3
djangorestframework==3.14.0
requests
gunicorn==20.1.0
django-cors-headers==3.13.0
celery==5.2.7
Redis==4.3.0